requirements.txt
anthropic>=0.40.0
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
rich>=13.7.0
numpy>=1.26.0
pytest>=8.0.0
supabase>=2.0.0
psycopg2-binary>=2.9.10
